backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
mysqlclient==2.2.4
pymongo==4.6.3
clickhouse-driver==0.2.7
python-dotenv==1.0.1
Pillow==10.3.0
django-filter==24.1
